processing/requirements.txt
numpy
Pillow
rawpy
PiDNG
opencv-python
pyautogui
pygetwindow
pytesseract